requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.13
django-bootstrap-v5==1.0.11
django-chartjs==2.3.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
